Updated the like icon. Now roadmaps can only be liked from the displayRoadmap view
