+++ accounts/templates/userProfile.html
@@ -70,9 +70,12 @@
                                     <div class="d-flex justify-content-end align-items-end mb-3 infoRoadmap2">
                                         <div class="d-flex flex-row align-items-center">
                                             <p style="margin: 0px 3px 0px 10px;">{{ roadmap.numberOfLikes }}</p>
-                                            <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-heart" viewBox="0 0 16 16">
-                                                <path d="m8 2.748-.717-.737C5.6.281 2.514.878 1.4 3.053c-.523 1.023-.641 2.5.314 4.385.92 1.815 2.834 3.989 6.286 6.357 3.452-2.368 5.365-4.542 6.286-6.357.955-1.886.838-3.362.314-4.385C13.486.878 10.4.28 8.717 2.01zM8 15C-7.333 4.868 3.279-3.04 7.824 1.143q.09.083.176.171a3 3 0 0 1 .176-.17C12.72-3.042 23.333 4.867 8 15"/>
-                                            </svg>
+                                            <div class="like d-flex justify-content-center">
+                                                <svg viewBox="0 0 256 256" height="20" width="20">
+                                                    <rect fill="none" height="16" width="16"></rect>
+                                                    <path d="M224.6,51.9a59.5,59.5,0,0,0-43-19.9,60.5,60.5,0,0,0-44,17.6L128,59.1l-7.5-7.4C97.2,28.3,59.2,26.3,35.9,47.4a59.9,59.9,0,0,0-2.3,87l83.1,83.1a15.9,15.9,0,0,0,22.6,0l81-81C243.7,113.2,245.6,75.2,224.6,51.9Z" stroke-width="20px" stroke="#000" fill="none"></path>
+                                                </svg>
+                                            </div>
                                         </div>
                                         <div class="d-flex flex-row align-items-center">
                                             <p style="margin: 0px 3px 0px 10px">{{ roadmap.completionPercentage }}%</p>
