+++ analytics/urls.py
@@ -4,5 +4,4 @@ from . import views
 urlpatterns = [
     path('', views.analytics, name='analytics'),
     path('explore', views.explore, name='explore'),
-    path('<int:roadmapID>/like', views.likeRoadmap, name='like'),
 ]
