+++ roadMap/static/styles/roadmapstyles.css
@@ -457,3 +457,81 @@ form div{
     rotate: 0;
 }
 }
+
+/* Like button */
+.containerLike input {
+    position: absolute;
+    opacity: 0;
+    cursor: pointer;
+    height: 0;
+    width: 0;
+  }
+  
+  .containerLike {
+    display: block;
+    position: relative;
+    cursor: pointer;
+    font-size: 20px;
+    user-select: none;
+    transition: 100ms;
+    margin : 10px 0;
+  }
+  
+  .checkmarkLike {
+    top: 0;
+    left: 0;
+    height: 2em;
+    width: 2em;
+    transition: 100ms;
+    animation: dislike_effect 400ms ease;
+  }
+
+  .checkmarkUnlike {
+    top: 0;
+    left: 0;
+    height: 2em;
+    width: 2em;
+    transition: 100ms;
+    animation: dislike_effect 400ms ease;
+  }
+  
+  .containerLike input:checked ~ .checkmarkLike path{
+    fill: #FF5353;
+    stroke-width: 0;
+  }
+  
+  .containerLike input:checked ~ .checkmarkLike {
+    animation: like_effect 400ms ease;
+  }
+  
+  .containerlike:hover {
+    transform: scale(1.1);
+  }
+  
+  @keyframes like_effect {
+    0% {
+      transform: scale(0);
+    }
+  
+    50% {
+      transform: scale(1.2);
+    }
+  
+    100% {
+      transform: scale(1);
+    }
+  }
+  
+  @keyframes dislike_effect {
+    0% {
+      transform: scale(0);
+    }
+  
+    50% {
+      transform: scale(1.2);
+    }
+  
+    100% {
+      transform: scale(1);
+    }
+  }
