+++ roadMap/templates/roadmap.html
@@ -83,6 +83,31 @@
             <div>
             </div>
         {% endfor %}
+        <!-- if the roadmap is liked by the user, show the button filled -->
+        {% if liked %}
+            <label class="containerLike d-flex align-items-center justify-content-center">
+                <a href="{% url 'like' roadmapId %}" class="containerlike">
+                    <input type="checkbox" checked="checked">
+                    <div class="checkmarkLike d-flex flex-row align-items-end">
+                        <svg viewBox="0 0 256 256">
+                            <rect fill="none" height="16" width="16"></rect>
+                            <path d="M224.6,51.9a59.5,59.5,0,0,0-43-19.9,60.5,60.5,0,0,0-44,17.6L128,59.1l-7.5-7.4C97.2,28.3,59.2,26.3,35.9,47.4a59.9,59.9,0,0,0-2.3,87l83.1,83.1a15.9,15.9,0,0,0,22.6,0l81-81C243.7,113.2,245.6,75.2,224.6,51.9Z" stroke-width="20px" stroke="#000" fill="none"></path></svg>
+                    </div>
+                </a>
+            </label>
+        {% else %}
+        <!-- if the roadmap is not liked by the user, show the button empty -->
+            <label class="containerLike d-flex align-items-center justify-content-center">
+                <a href="{% url 'like' roadmapId %}" class="containerlike">
+                    <input type="checkbox" checked="checked">
+                    <div class="checkmarkUnlike d-flex flex-row align-items-end">
+                        <svg viewBox="0 0 256 256">
+                            <rect fill="none" height="16" width="16"></rect>
+                            <path d="M224.6,51.9a59.5,59.5,0,0,0-43-19.9,60.5,60.5,0,0,0-44,17.6L128,59.1l-7.5-7.4C97.2,28.3,59.2,26.3,35.9,47.4a59.9,59.9,0,0,0-2.3,87l83.1,83.1a15.9,15.9,0,0,0,22.6,0l81-81C243.7,113.2,245.6,75.2,224.6,51.9Z" stroke-width="20px" stroke="#000" fill="none"></path></svg>
+                    </div>
+                </a>
+            </label>
+        {% endif %}
     <div class="footer">
         <h2>Conclusions</h2>
         <div class="contenedorConclutions">
