+++ roadMap/urls.py
@@ -7,4 +7,5 @@ urlpatterns = [
     path('checkpointUpdate', views.checkpointUpdate, name='checkpointUpdate'),
     path('displayRoadmap/<int:roadmapId>', views.displayRoadmap, name='displayRoadmap'),
     path('displayRoadmap/<int:roadmapId>/<int:stepNumber>', views.displayRoadmap, name='displayRoadmap'),
+    path('displayRoadmap/<int:roadmapID>/like', views.likeRoadmap, name='like'),
 ]
